+++ react/WeaponRow.js
@@ -14,7 +14,10 @@ class WeaponRow extends React.Component {
         this.readiedBaseI = -3;
         this.baseCheckBonusForSlowActions = 5;
         this.extraActionModifier = 10;
-        this.penaltyCounterStat = "INT";
+    }
+
+    penaltyCounterStat() {
+        return "INT"
     }
 
     skillLevel() {
@@ -44,14 +47,16 @@ class WeaponRow extends React.Component {
     }
 
     baseROA() {
-        return parseFloat(this.props.weapon.base.roa) +
-            (-0.15) * (this.props.weapon.size - 1) +
-            parseFloat(this.props.weapon.quality.roa);
+        const bd = new ValueBreakdown()
+        bd.add(parseFloat(this.props.weapon.base.roa), "Base")
+        bd.add((-0.15) * (this.props.weapon.size - 1), "size")
+        bd.add(parseFloat(this.props.weapon.quality.roa), "quality")
+        return bd
     }
 
-    skillROAMultiplier() {
+    skillROAMultiplier(skillName) {
         const level = this.props.skillHandler.skillLevel(
-            this.props.weapon.base.base_skill.name);
+            skillName ?? this.props.weapon.base.base_skill.name);
         if (level > 0) {
             return (1 + 0.1 * level);
         }
@@ -62,7 +67,8 @@ class WeaponRow extends React.Component {
         if (!useType) {
             useType = WeaponRow.FULL;
         }
-        let roa = this.baseROA();
+
+        const roa = this.baseROA();
 
         let specLevel;
         if (useType === WeaponRow.SPECIAL || useType === WeaponRow.FULL) {
@@ -71,25 +77,20 @@ class WeaponRow extends React.Component {
             if (!util.isInt(specLevel)) {
                 specLevel = 0;
             }
-            roa += specLevel * 0.05;
+            roa.add(specLevel * 0.05, "SWS");
         } else {
             specLevel = this.props.skillHandler.skillLevel(
                 "Two-weapon style");
             if (!util.isInt(specLevel)) {
                 specLevel = 0;
             }
-            let mod;
-            if (useType === WeaponRow.PRI) {
-                mod = -0.25;
-            } else if (useType === WeaponRow.SEC) {
-                mod = -0.5;
-            }
-            mod += specLevel * 0.05;
-
-            roa += Math.min(mod, 0);
+            const fromUseType = useType === WeaponRow.PRI ? -0.25 : -0.5
+            roa.add(fromUseType, "use type")
+            roa.add(Math.min(specLevel * 0.05, -fromUseType), "counter from TWS")
         }
-        roa *= this.skillROAMultiplier();
-        return {value: Math.min(roa, 2.5), breakdown: []};
+        roa.multiply(this.skillROAMultiplier(), "from skill");
+        roa.setMaximum(2.5, "Max ROA")
+        return roa;
     }
 
     ccv() {
@@ -174,11 +175,11 @@ class WeaponRow extends React.Component {
             props = Object.assign(props, givenProps);
         }
 
-        const roa = this.roa(props.useType).value;
+        const roa = this.roa(props.useType).value();
         const baseCheck = this.skillCheck();
         if (!baseCheck) {
             // Actions not available.
-            return null; //actions.map((e) => {return null;});
+            return null;
         }
         const checks = [];
 
@@ -213,7 +214,7 @@ class WeaponRow extends React.Component {
                 // TODO: counterPenalty is a bad name, as a bad stat will give actual penalty for actions with these, see AE2K_Weapons_17.xls
                 if (props.counterPenalty) {
                     let counter = WeaponRow.counterPenaltyV2(
-                        this.getStat(this.penaltyCounterStat))
+                        this.getStat(this.penaltyCounterStat()))
                     if (counter > 0) {
                         if (mod > 0) {
                             counter = 0
@@ -222,7 +223,7 @@ class WeaponRow extends React.Component {
                         }
                     }
 
-                    bd.add(counter, `Modifier from ${this.penaltyCounterStat}`)
+                    bd.add(counter, `Modifier from ${this.penaltyCounterStat()}`)
                 }
                 checks.push(bd)
             }
@@ -261,7 +262,7 @@ class WeaponRow extends React.Component {
         if (givenProps) {
             props = Object.assign(props, givenProps);
         }
-        const rof = this.roa(props.useType).value;
+        const rof = this.roa(props.useType).value();
         const baseI = -5 / rof;
         const readiedBaseI = this.readiedBaseI;
         let targetI = this.targetInitiative();
@@ -444,7 +445,7 @@ class WeaponRow extends React.Component {
             return <td key={`di-${ii}`} style={defenseInitStyle} aria-label={`Defense initiative for ${useType}`}
             >{this.renderInt(el)}</td>;});
 
-        return <tr aria-label={`Action row for ${useType}`}><td style={cellStyle} aria-label={`ROA for ${useType}`}>{this.roa(useType).value.toFixed(2)}</td>
+        return <tr aria-label={`Action row for ${useType}`}><td style={cellStyle} aria-label={`ROA for ${useType}`}>{this.roa(useType).value().toFixed(2)}</td>
             {checkCells}
             {attackInitiativeCells}
             <td style={attackDamageStyle} aria-label={`Attack damage for ${useType}`}>{
@@ -503,7 +504,6 @@ class WeaponRow extends React.Component {
             <span style={infoStyle}><label>Bypass:</label> <span aria-label={"Bypass"}>{this.bypass()}</span></span>
             <span style={infoStyle}><label>Weight:</label> <span aria-label={"Weight"}>{util.itemWeight(this.props.weapon).toFixed(2)} kg</span></span>
                     <Button onClick={(e) => this.handleRemove()}
-                            ref={(c) => this._removeButton = c}
                             size="sm"
                     >Remove</Button>
             </div>